moving service-card styles to a file
